fix(daemon): use ru_maxrss directly without conversion
The conversion was redundant on all targets we build on: c_long is i64
on 64-bit Linux/macOS (no conversion needed) and i32 on 32-bit Linux
(the subsequent 'as u64' sign-extends to i64 first, then saturating_mul
operates on u64). Removing the conversion silences both
clippy::unnecessary_cast and clippy::useless_conversion without losing
overflow safety.

+++ crates/daemon/tests/connection_scaling_stress.rs
@@ -247,7 +247,7 @@ fn current_rss_bytes() -> Option<u64> {
     if rc != 0 {
         return None;
     }
-    let raw = i64::from(usage.ru_maxrss);
+    let raw = usage.ru_maxrss;
     if raw <= 0 {
         return None;
     }
